fix(web): expose Function creation from the dashboard New button
The dashboard's "New Workflow" button on the workflow-list section
went straight to NEW_WORKFLOW_NAV — Function creation was reachable
only from the /workflows index page. Two surfaces, two different
discoverability stories.

Extract the existing dropdown JSX (~30 lines: Workflow + Function
items with descriptions and the Beta badge) into a shared
NewArtifactDropdown component and use it from both the dashboard
workflow-list header and the workflows index page header. The
empty-state CTAs on both pages remain "New Workflow" single buttons
— matches the existing pattern and avoids surfacing Function
creation when the user has zero artifacts of any kind.
